requirements.txt
pandas
numpy
requests
ta
yfinance
openai
matplotlib
finnhub-python>=1.9.0